Reorganize train_track.py
TrainTrackSelfMap class removed. Instead of deriving a separate class for self
maps from TrainTrackMap, it seems simpler to just use TrainTrackMap for this,
too. Its train track specific methods on homology computations are moved to
TrainTrack and train track map specific methods are moved to TrainTrackMap.

MeasuredTrainTrack also removed. Measured train tracks will instead be
represented as TrainTracks with a measure attribute.

Splitting, shifting and folding-related methods are sketched to TrainTrack. Since it is much faster to do a split by changing the train track instead of creating a new copy, it makes sense that such methods are part of the TrainTrack class.